On the way to more compact projects

+++ solver/project/src/fitness.rs
@@ -4,7 +4,7 @@ use crate::Schedule;
 /// the makespan but in future, this could be extended penalise
 /// infeasibility (e.g. resource use conflicts) or to take job
 /// deadlines into account.
-pub fn fitness(schedule: &Schedule) -> Vec<i64> {
+pub fn fitness(schedule: &Schedule) -> Vec<f64> {
     let makespan = schedule.makespan();
     // For now, return makespan as the only component of the fitness function
     vec![makespan]
